+++ docs/src/release_notes.rst
@@ -9,6 +9,7 @@ Release Notes
 0.8.27
 ------
 + Added decorator to provide common methods for Playbook Apps.
++ Added :py:meth:`~tcex.tcex_playbook.TcExPlaybook.add_output` and :py:meth:`~tcex.tcex_playbook.TcExPlaybook.write_output` methods to provide an alternative way to write playbook output data.
 + Updated ``tclib`` logic for **lib_latest** symbolic link.
 
 0.8.26